requirements.txt
gradio
groq
python-dotenv
streamlit
pillow
requests